IC-0000 PR Feedback - moved community-api/delius specific model objects to a separate package to crs
