+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsdeliusinterventionseventlistener/model/communityapi/Contact.kt
@@ -1,0 +1,5 @@
+package uk.gov.justice.digital.hmpps.hmppsdeliusinterventionseventlistener.model.communityapi
+
+data class Contact(
+  val contactId: Long
+)
