--- src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsdeliusinterventionseventlistener/model/crs/Contact.kt
@@ -1,5 +1,0 @@
-package uk.gov.justice.digital.hmpps.hmppsdeliusinterventionseventlistener.model.crs
-
-data class Contact(
-  val contactId: Long
-)
